ALPHA (V06-11-23)
I've been going for the last couple hours now trying to get design elements to behave properly and it (kinda) works now... 

I've got a new 'Home' label on the top and the new background design. I don't mind the background; a couple of things need to be changed but NOT sure if the 'title label' at the top has sold me. It looks slightly out of place. 

Things have been centred but CSS still takes the crown of how much code. I want to make this website really fluid and reactant based, not a huge fan of static elements. 

My priorities in the next couple of days will 100% be stability. It's currently really good across mobile devices bigger than 5.5 inches which I'm happy about.

Signing off for the night

- Alic - ScytheDEM